@@ -2,6 +2,8 @@ package socialnet.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -10,14 +12,21 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 import socialnet.api.request.LoginRq;
-import socialnet.api.response.*;
+import socialnet.api.request.UserRq;
+import socialnet.api.request.UserUpdateDto;
+import socialnet.api.response.CommonRs;
+import socialnet.api.response.ComplexRs;
+import socialnet.api.response.ErrorRs;
+import socialnet.api.response.PersonRs;
 import socialnet.exception.EmptyEmailException;
+import socialnet.mappers.PersonMapper;
+import socialnet.mappers.UserDtoMapper;
 import socialnet.model.Person;
 import socialnet.repository.PersonRepository;
 import socialnet.security.jwt.JwtUtils;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import java.time.LocalDate;
+import java.util.ResourceBundle;
 
 
 @Service
@@ -29,8 +38,12 @@ public class PersonService {
     private String jwt;
     private final AuthenticationManager authenticationManager;
     private final PersonRepository personRepository;
-    private static final SimpleDateFormat DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
+    private final WeatherService weatherService;
+    private final CurrencyService currencyService;
+    private final PersonMapper personMapper;
+    private final UserDtoMapper userDtoMapper;
 
+    private static final ResourceBundle textProperties = ResourceBundle.getBundle("text");
 
     public Object getLogin(LoginRq loginRq) {
 
@@ -38,18 +51,58 @@ public class PersonService {
         if ((person = checkLoginAndPassword(loginRq.getEmail(), loginRq.getPassword())) != null) {
             jwt = jwtUtils.generateJwtToken(loginRq.getEmail());
             authenticated(loginRq);
-
-            return setLoginRs(jwt, person);
+            PersonRs personRs = personMapper.toDTO(person);
+            changePersonStatus(personRs);
+            return new CommonRs<>(personRs);
         } else {
             throw new EmptyEmailException("invalid username or password");
         }
     }
 
-    public Object getMe(String authorization) {
-
+    public CommonRs<PersonRs> getMyProfile(String authorization) {
         String email = jwtUtils.getUserEmail(authorization);
         Person person = personRepository.findByEmail(email);
-        return setLoginRs(jwt, person);
+        PersonRs personRs = personMapper.toDTO(person);
+        changePersonStatus(personRs);
+        return new CommonRs<>(personRs);
+    }
+
+    private void changePersonStatus(PersonRs personRs) {
+        personRs.setToken(jwt);
+        personRs.setOnline(true);
+        personRs.setIsBlockedByCurrentUser(false);
+        personRs.setIsBlockedByCurrentUser(false);
+        personRs.setWeather(weatherService.getWeatherByCity(personRs.getCity()));
+        personRs.setCurrency(currencyService.getCurrency(LocalDate.now()));
+        if (personRs.getPhoto() == null) {
+            personRs.setPhoto(textProperties.getString("default.photo"));
+        }
+
+    }
+
+    public ResponseEntity<?> getUserInfo(String authorization) {
+
+        if (!jwtUtils.validateJwtToken(authorization)) {//401
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
+        String userName = jwtUtils.getUserEmail(authorization);
+        if (userName.isEmpty()) {
+            return new ResponseEntity<>(
+                    new ErrorRs("EmptyEmailException","Field 'email' is empty"), HttpStatus.BAD_REQUEST);  //400
+        }
+
+        Person person = personRepository.findByEmail(userName);
+        if (person.getIsDeleted()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);  //403
+        }
+
+        PersonRs personRs = personMapper.toDTO(person);
+
+        personRs.setWeather(weatherService.getWeatherByCity(person.getCity()));
+        personRs.setCurrency(currencyService.getCurrency(LocalDate.now()));
+
+        return ResponseEntity.ok(new CommonRs(personRs));
     }
 
     public Object getLogout(String authorization) {
@@ -59,8 +112,9 @@ public class PersonService {
 
     public Object getUserById(String authorization, Integer id) {
         Person person = findUser(id);
-        log.info("вход на страницу пользователя " + person.getFirstName());
-        return setLoginRs(authorization, person);
+        PersonRs personRs = personMapper.toDTO(person);
+        changePersonStatus(personRs);
+        return new CommonRs<>(personRs);
     }
 
     private Person findUser(Integer id) {
@@ -88,61 +142,16 @@ public class PersonService {
                 .build();
     }
 
-    public WeatherRs setLoginWeather(WeatherRs weatherRs, Person person) {
-
-        weatherRs.setCity(person.getCity());
-        weatherRs.setTemp("");
-        weatherRs.setDate(DATE_FORMATTER.format(new Date()));
-        weatherRs.setClouds("");
-
-        return weatherRs;
+    public Person getAuthPerson() {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        return personRepository.getPersonByEmail(email);
     }
 
-    public CurrencyRs setCurrencyRs(CurrencyRs currencyRs) {
-
-        currencyRs.setEuro("");
-        currencyRs.setUsd("");
-
-        return currencyRs;
+    public Long getAuthPersonId() {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        return personRepository.getPersonIdByEmail(email);
     }
 
-    public PersonRs setPersonRs(CurrencyRs currencyRs, WeatherRs weatherRs, String jwt, Person person) {
-        PersonRs personRs = new PersonRs();
-        weatherRs = setLoginWeather(weatherRs, person);
-        currencyRs = setCurrencyRs(currencyRs);
-        personRs.setAbout(person.getAbout());
-        personRs.setCity(person.getCity());
-        personRs.setCountry(person.getCountry());
-        personRs.setBirthDate(person.getBirthDate());
-        personRs.setCurrency(currencyRs);
-        personRs.setWeather(weatherRs);
-        personRs.setEmail(person.getEmail());
-        personRs.setFirstName(person.getFirstName());
-        personRs.setFriendStatus("");
-        personRs.setId(person.getId());
-        personRs.setIsBlocked(person.getIsBlocked());
-        personRs.setIsBlockedByCurrentUser(false);
-        personRs.setLastName(person.getLastName());
-        personRs.setLastOnlineTime(person.getLastOnlineTime());
-        personRs.setMessagesPermission(person.getMessagePermissions());
-        personRs.setOnline(true);
-        personRs.setPhone(person.getPhone());
-        personRs.setPhoto(person.getPhoto());
-        personRs.setRegDate(person.getRegDate());
-        personRs.setToken(jwt);
-        personRs.setUserDeleted(person.getIsDeleted());
-        return personRs;
-    }
-
-    public CommonRs<PersonRs> setLoginRs(String jwt, Person person) {
-        WeatherRs loginWeather = new WeatherRs();
-        CurrencyRs currencyRs = new CurrencyRs();
-        PersonRs personRs = setPersonRs(currencyRs, loginWeather, jwt, person);
-        CommonRs<PersonRs> commonRs = new CommonRs<>();
-        commonRs.setData(personRs);
-        commonRs.setTimestamp(System.currentTimeMillis());
-        return commonRs;
-    }
 
     public Person checkLoginAndPassword(String email, String password) {
 
@@ -161,5 +170,24 @@ public class PersonService {
         SecurityContextHolder.getContext().setAuthentication(authentication);
     }
 
+    public ResponseEntity<?> updateUserInfo(String authorization, UserRq userRq) {
+
+        Person person = personRepository.findByEmail(jwtUtils.getUserEmail(authorization));
+
+        if (person.getIsBlocked()) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);  //403
+        }
+
+        PersonRs personRs = personMapper.toDTO(person);
+        UserUpdateDto userUpdateDto = userDtoMapper.toDto(userRq);
+
+        userUpdateDto.setPhoto(person.getPhoto());
+        if (userUpdateDto.getPhoto() == null)
+            userUpdateDto.setPhoto(textProperties.getString("default.photo"));
+
+        personRepository.updatePersonInfo(userUpdateDto, person.getEmail());
+
+        return ResponseEntity.ok(new CommonRs(personRs));
+    }
 
 }
