@@ -1,17 +1,17 @@
 package socialnet.api.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@AllArgsConstructor
+@NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class WeatherRs {
-
     private String city;
-
     private String clouds;
-
     private String date;
-
     private String temp;
 }
