@@ -1,26 +1,21 @@
 package socialnet.schedules;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+import socialnet.repository.CaptchaRepository;
 
 @Component
 @EnableAsync
 @RequiredArgsConstructor
 public class RemoveOldCaptchasSchedule {
-    private final JdbcTemplate jdbcTemplate;
+    private final CaptchaRepository captchaRepository;
 
-    // Выполнять команду каждый день в 00:00
     @Async
-    @Scheduled(cron = "0 0 0 * * *")
+    @Scheduled(cron = "${schedules.deleteOldCaptchasInterval}")
     public void removeOldCaptchas() {
-        try {
-            jdbcTemplate.update("DELETE FROM captcha WHERE time < now() - interval '1 day'");
-        } catch (DataAccessException ignored) {
-        }
+        captchaRepository.removeOldCaptchas();
     }
 }
