@@ -1,16 +1,20 @@
 package socialnet.utils;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import socialnet.repository.PersonRepository;
 
 import java.sql.Timestamp;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.ResourceBundle;
 
 @Component
+@RequiredArgsConstructor
 public class Converter {
-    private final String ISO_DATE_FORMAT_ZERO_OFFSET = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
-    private final String UTC_TIMEZONE_NAME = "UTC";
+    private static final ResourceBundle textProperties = ResourceBundle.getBundle("text");
+    private final PersonRepository personRepository;
 
     public Long dateToMillisec(String dateStr
     ) throws ParseException {
@@ -22,6 +26,14 @@ public class Converter {
     public Timestamp dateToTimeStamp(String dateStr) {
         dateStr = dateStr.substring(0,19).replace("T"," ");
         return Timestamp.valueOf(dateStr);
+    }
+
+    public String checkPhotoId(String photoId) {
+        if (photoId == null)
+            return textProperties.getString("default.photo");
+        else
+            return photoId;
+
     }
 
 }
